fix: MyBook API updateProcess 에러 해결

+++ src/main/java/megabooks/megabooks/domain/myBook/entity/MyBook.java
@@ -28,6 +28,6 @@ public class MyBook extends BaseEntity {
     private Book book;
 
     public void updateMyBookProcess(MyBookRequestDTO.MyBookProcessDTO myBookProcessDTO) {
-        this.myBookProcess = (myBookProcessDTO.getNowPage()/myBookProcessDTO.getTotalPage())*100;
+        this.myBookProcess = ((double) myBookProcessDTO.getNowPage() / myBookProcessDTO.getTotalPage()) * 100;
     }
 }
